% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_correction2.R
\name{count_correction2}
\alias{count_correction2}
\title{A helper function to generate effort (detectability) corrected counts for each checklist}
\usage{
count_correction2(sp_data)
}
\arguments{
\item{sp_data}{zero-filled 'sp_data' data.table object, output from \code{import_from_erd} function}
}
\value{
A \code{data.table} object with zero-filled abundance (effort corrected) and presence-only reported colums
as well as additional columns for latitude, longitude, year, month, day_of_year, hours_of_day,
protocol_id, is_stationary, is_traveling, effort_hrs, effort_distance_km, cci
}
\description{
This function filters non-zero counts and then runs XGboost to predict count for
original conditions (checlist duration, distance, hour of day, number of observer, mode of checklist,
cci), as well for standarndized condition (medians). Then, it estimates the 'correction factor, cf'
which is used to standardize (correct for effort and detection) counts.
}
