% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacency_matrix.R
\name{get_adjacency_matrix}
\alias{get_adjacency_matrix}
\title{Get adjacency matrix}
\usage{
get_adjacency_matrix(data, grid_large)
}
\arguments{
\item{data}{a vector of cell indices}

\item{grid}{a ddgridR grid}

\item{n_small_min}{minimum number of small cells required to calculate an abundance ratio for a large cell}

\item{quiet}{if TRUE, suppress informational messages (warnings/errors still returned)}
}
\value{
a square matrix indicating which cells are direct neighbors (value 1) and which cells are not (value 0)
}
\description{
Get adjacency matrix
}
