% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_erd.R
\name{import_checklists}
\alias{import_checklists}
\title{Import checklists data from the erd}
\usage{
import_checklists(
  checklists_parquet_path =
    "~/Documents/macrodemo_project/input_data/ebird2022/checklists-2022.parquet",
  cci_weather_path =
    "~/Documents/macrodemo_project/input_data/ebird2022/erd2022-cci-weather.parquet"
)
}
\arguments{
\item{erd_path}{The path to the eBird reference dataset .db file}
}
\value{
a data.frame of the checklists
}
\description{
Import checklists data from the erd
}
